Correct sort.h

+++ sort.h
@@ -20,7 +20,7 @@ typedef struct listint_s
 void print_array(const int *array, size_t size);
 void print_list(const listint_t *list);
 void bubble_sort(int *array, size_t size);
-void swap_elements(int *array, size_t i, size_t j);
+void swap_elements(int *array, int i, int j);
 void insertion_sort_list(listint_t **list);
 void selection_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
